feat(updated): add useBookMarks & useBookmarksList hooks
