+++ src/screens/Bookmarks/Bookmarks.tsx
@@ -1,42 +1,19 @@
-import React, {useCallback} from 'react';
+import React from 'react';
 import {View, ScrollView} from 'react-native';
 import {styles} from './styles';
 import {BookmarkItem, SuspenseView} from 'atoms';
-import {selectBookmarksCardsData} from 'core/selectors';
-import {useSelector, useDispatch} from 'react-redux';
-import {
-  useRequestError,
-  useRequestLoading,
-  useBookmarksAnalytics,
-} from 'core/hooks';
-import {getInitialHomeDataRequest} from 'core/reducers';
-import {IProps} from './types';
-import {IBookmarkItem} from 'core/types';
 import {isEmpty} from 'lodash';
 import {BookmarksEmptyView} from 'molecules';
+import {useBookmarks} from './hooks';
 
-export const Bookmarks = ({navigation}: IProps) => {
-  const bookmarksCategories = useSelector(selectBookmarksCardsData);
-  const dispatch = useDispatch();
-  const {sendSelectSavedCategoryEvent} = useBookmarksAnalytics();
-
-  const getHomeData = useCallback(() => {
-    dispatch(getInitialHomeDataRequest());
-  }, [dispatch]);
-
-  const {loading} = useRequestLoading(getInitialHomeDataRequest);
-  const {error} = useRequestError(getInitialHomeDataRequest);
-
-  const navigateToBookmarksList = useCallback(
-    ({categoryName, categoryId}: IBookmarkItem) => {
-      navigation.navigate('BookmarksList', {
-        title: categoryName,
-        categoryId: categoryId,
-      });
-      sendSelectSavedCategoryEvent(categoryName);
-    },
-    [navigation, sendSelectSavedCategoryEvent],
-  );
+export const Bookmarks = () => {
+  const {
+    getHomeData,
+    bookmarksCategories,
+    loading,
+    error,
+    navigateToBookmarksList,
+  } = useBookmarks();
 
   return (
     <SuspenseView
