+++ src/screens/Bookmarks/hooks/index.tsx
@@ -1,0 +1,1 @@
+export {useBookmarks} from './useBookmarks';
