+++ src/screens/BookmarksList/BookmarksList.tsx
@@ -2,34 +2,22 @@ import React, {useCallback, useEffect, useLayoutEffect, useMemo} from 'react';
 import {FlatList} from 'react-native';
 import {ObjectCard} from 'molecules';
 import {styles} from './styles';
-import {IProps} from './types';
 
 import {isAndroid, isIOS, SCREEN_WIDTH} from 'services/PlatformService';
 import {PADDING_HORIZONTAL} from 'core/constants';
-import {useBookmarksObjects, useBookmarksListAnalytics} from 'core/hooks';
-import {IObject} from 'core/types';
-import {orderBy} from 'lodash';
+import {useBookmarksList} from './hooks';
 
 const cardWidth = SCREEN_WIDTH - PADDING_HORIZONTAL * 2;
 
-export const BookmarksList = ({
-  route,
-  navigation: {setOptions, goBack, navigate},
-}: IProps) => {
+export const BookmarksList = () => {
   const {
-    params: {title, categoryId},
-  } = route;
-
-  const listData = useBookmarksObjects(categoryId);
-
-  const {sendSaveCardEvent, sendSelectCardEvent, sendUnsaveCardEvent} =
-    useBookmarksListAnalytics();
-
-  const sortedListData = useMemo(() => {
-    return listData
-      ? orderBy(listData, [({name}) => name.toLowerCase()], 'asc')
-      : null;
-  }, [listData]);
+    setOptions,
+    sortedListData,
+    goBack,
+    title,
+    navigateToObjectDetails,
+    sendIsFavoriteChangedEvent,
+  } = useBookmarksList();
 
   useLayoutEffect(() => {
     setOptions({
@@ -48,25 +36,6 @@ export const BookmarksList = ({
       goBack();
     }
   }, [goBack, onLastObjectRemoveAnimationEnd, sortedListData]);
-
-  const navigateToObjectDetails = useCallback(
-    ({id, name, category}: IObject) => {
-      navigate('ObjectDetails', {objectId: id});
-      sendSelectCardEvent(name, category.name);
-    },
-    [navigate, sendSelectCardEvent],
-  );
-
-  const sendIsFavoriteChangedEvent = useCallback(
-    ({name, category}: IObject, nextIsFavoriteStatus: boolean) => {
-      if (nextIsFavoriteStatus) {
-        sendSaveCardEvent(name, category.name);
-      } else {
-        sendUnsaveCardEvent(name, category.name);
-      }
-    },
-    [sendSaveCardEvent, sendUnsaveCardEvent],
-  );
 
   return (
     <FlatList
