+++ src/screens/BookmarksList/hooks/index.tsx
@@ -1,0 +1,1 @@
+export {useBookmarksList} from './useBookmarksList'
